backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Authenticated Todo Application Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlmodel==0.0.14",
    "psycopg2-binary==2.9.9",
    "alembic==1.12.0",
    "pydantic==2.5.0",
    "python-jose[cryptography]==3.3.0",
    "python-dotenv==1.0.0",
    "pydantic-settings==2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
